Reorganize acts-related code into dedicated directory
