+++ packages/giselle-engine/src/core/acts/get-workspace-acts.ts
@@ -1,8 +1,8 @@
 import type { WorkspaceId } from "@giselle-sdk/data-type";
 import type { GiselleEngineContext } from "../types";
 import { getWorkspaceIndex } from "../utils/workspace-index";
-import { ActIndexObject, ActObject } from "./act/object";
-import { actPath, workspaceActPath } from "./act/paths";
+import { ActIndexObject, ActObject } from "./object";
+import { actPath, workspaceActPath } from "./object/paths";
 
 export async function getWorkspaceActs(args: {
 	context: GiselleEngineContext;
